load all data in navbar, take all ajax gets from everywhere
