+++ public/js/application.js
@@ -54,7 +54,7 @@
 /******/ 	
 /******/ 	
 /******/ 	var hotApplyOnUpdate = true;
-/******/ 	var hotCurrentHash = "0871a153dcc80d1fe122"; // eslint-disable-line no-unused-vars
+/******/ 	var hotCurrentHash = "8ad1f99ad320ad61c316"; // eslint-disable-line no-unused-vars
 /******/ 	var hotCurrentModuleData = {};
 /******/ 	var hotCurrentParents = []; // eslint-disable-line no-unused-vars
 /******/ 	
@@ -1994,13 +1994,13 @@
 /* 241 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.DigestContainer = exports.UpdateContainer = exports.UpdateList = exports.Update = exports.SectionContainer = exports.Digest = exports.Editor = exports.DigestList = exports.Navbar = undefined;\n\nvar _navbar = __webpack_require__(242);\n\nvar _navbar2 = _interopRequireDefault(_navbar);\n\nvar _digestList = __webpack_require__(295);\n\nvar _digestList2 = _interopRequireDefault(_digestList);\n\nvar _editor = __webpack_require__(300);\n\nvar _editor2 = _interopRequireDefault(_editor);\n\nvar _digest = __webpack_require__(301);\n\nvar _digest2 = _interopRequireDefault(_digest);\n\nvar _sectionContainer = __webpack_require__(302);\n\nvar _sectionContainer2 = _interopRequireDefault(_sectionContainer);\n\nvar _update = __webpack_require__(304);\n\nvar _update2 = _interopRequireDefault(_update);\n\nvar _updateList = __webpack_require__(305);\n\nvar _updateList2 = _interopRequireDefault(_updateList);\n\nvar _updateContainer = __webpack_require__(307);\n\nvar _updateContainer2 = _interopRequireDefault(_updateContainer);\n\nvar _digestContainer = __webpack_require__(308);\n\nvar _digestContainer2 = _interopRequireDefault(_digestContainer);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nexports.Navbar = _navbar2.default;\nexports.DigestList = _digestList2.default;\nexports.Editor = _editor2.default;\nexports.Digest = _digest2.default;\nexports.SectionContainer = _sectionContainer2.default;\nexports.Update = _update2.default;\nexports.UpdateList = _updateList2.default;\nexports.UpdateContainer = _updateContainer2.default;\nexports.DigestContainer = _digestContainer2.default;\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"index.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/index.js\n ** module id = 241\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/index.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.DigestContainer = exports.UpdateContainer = exports.UpdateList = exports.Update = exports.SectionContainer = exports.Digest = exports.Editor = exports.DigestList = exports.Navbar = undefined;\n\nvar _navbar = __webpack_require__(242);\n\nvar _navbar2 = _interopRequireDefault(_navbar);\n\nvar _digestList = __webpack_require__(299);\n\nvar _digestList2 = _interopRequireDefault(_digestList);\n\nvar _editor = __webpack_require__(301);\n\nvar _editor2 = _interopRequireDefault(_editor);\n\nvar _digest = __webpack_require__(302);\n\nvar _digest2 = _interopRequireDefault(_digest);\n\nvar _sectionContainer = __webpack_require__(303);\n\nvar _sectionContainer2 = _interopRequireDefault(_sectionContainer);\n\nvar _update = __webpack_require__(305);\n\nvar _update2 = _interopRequireDefault(_update);\n\nvar _updateList = __webpack_require__(306);\n\nvar _updateList2 = _interopRequireDefault(_updateList);\n\nvar _updateContainer = __webpack_require__(307);\n\nvar _updateContainer2 = _interopRequireDefault(_updateContainer);\n\nvar _digestContainer = __webpack_require__(308);\n\nvar _digestContainer2 = _interopRequireDefault(_digestContainer);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nexports.Navbar = _navbar2.default;\nexports.DigestList = _digestList2.default;\nexports.Editor = _editor2.default;\nexports.Digest = _digest2.default;\nexports.SectionContainer = _sectionContainer2.default;\nexports.Update = _update2.default;\nexports.UpdateList = _updateList2.default;\nexports.UpdateContainer = _updateContainer2.default;\nexports.DigestContainer = _digestContainer2.default;\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"index.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/index.js\n ** module id = 241\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/index.js?");
 
 /***/ },
 /* 242 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(180);\n\nvar _reduxSimpleRouter = __webpack_require__(240);\n\nvar _reactMdl = __webpack_require__(243);\n\nvar _account = __webpack_require__(290);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Navbar = (function (_React$Component) {\n  _inherits(Navbar, _React$Component);\n\n  function Navbar() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, Navbar);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(Navbar)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.fetchAccountInfo = function () {\n      var ajax = new XMLHttpRequest();\n      ajax.open('GET', '/api/account');\n      ajax.onreadystatechange = function () {\n        if (ajax.readyState != XMLHttpRequest.DONE) {\n          return;\n        }\n        if (ajax.status !== 200) {\n          _this.props.dispatch((0, _account.loadError)());\n        }\n        var payload = JSON.parse(ajax.response);\n        _this.props.dispatch((0, _account.loadSuccess)(payload.data));\n      };\n      ajax.send();\n      _this.props.dispatch((0, _account.loadStart)());\n    }, _this.componentDidMount = function () {\n      _this.fetchAccountInfo();\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(Navbar, [{\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      if (this.props.currentUser.get('loading')) {\n        return _react2.default.createElement(\n          'h1',\n          null,\n          'Loading...'\n        );\n      }\n\n      return _react2.default.createElement(\n        _reactMdl.Layout,\n        { fixedHeader: true, fixedDrawer: true },\n        _react2.default.createElement(\n          _reactMdl.Drawer,\n          { title: 'Kindred' },\n          _react2.default.createElement('img', { src: this.props.currentUser.get('avatarUrl'), height: '240px' }),\n          _react2.default.createElement(\n            'p',\n            null,\n            this.props.currentUser.get('firstName') + ' ' + this.props.currentUser.get('lastName')\n          ),\n          _react2.default.createElement(\n            _reactMdl.Navigation,\n            null,\n            _react2.default.createElement(\n              'a',\n              { onClick: function onClick() {\n                  _this2.props.dispatch((0, _reduxSimpleRouter.pushPath)('/'));\n                } },\n              'Dashboard'\n            ),\n            _react2.default.createElement(\n              'a',\n              { onClick: function onClick() {\n                  _this2.props.dispatch((0, _reduxSimpleRouter.pushPath)('/editor'));\n                } },\n              'Editor'\n            ),\n            _react2.default.createElement(\n              'a',\n              { onClick: function onClick() {\n                  _this2.props.dispatch((0, _reduxSimpleRouter.pushPath)('/updates'));\n                } },\n              'My Updates'\n            )\n          )\n        ),\n        _react2.default.createElement(\n          _reactMdl.Header,\n          { title: 'Kindred' },\n          _react2.default.createElement(_reactMdl.Textfield, {\n            label: 'Search',\n            expandable: true,\n            expandableIcon: 'search'\n          })\n        ),\n        _react2.default.createElement(\n          'main',\n          { className: 'mdl-layout__content' },\n          _react2.default.createElement(\n            'div',\n            { className: 'page-content' },\n            this.props.children\n          )\n        )\n      );\n    }\n  }]);\n\n  return Navbar;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    currentUser: state.currentUser,\n    loading: state.content.loading\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(Navbar);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"navbar.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/navbar.js\n ** module id = 242\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/navbar.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactRedux = __webpack_require__(180);\n\nvar _reduxSimpleRouter = __webpack_require__(240);\n\nvar _reactMdl = __webpack_require__(243);\n\nvar _account = __webpack_require__(290);\n\nvar _digest = __webpack_require__(295);\n\nvar _section = __webpack_require__(296);\n\nvar _update = __webpack_require__(297);\n\nvar _myUpdate = __webpack_require__(298);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Navbar = (function (_React$Component) {\n  _inherits(Navbar, _React$Component);\n\n  function Navbar() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, Navbar);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(Navbar)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.componentDidMount = function () {\n      _this.fetchAccountInfo();\n      _this.fetchAllDigests();\n      _this.fetchAllUpdates();\n    }, _this.fetchAccountInfo = function () {\n      var ajax = new XMLHttpRequest();\n      ajax.open('GET', '/api/account');\n      ajax.onreadystatechange = function () {\n        if (ajax.readyState != XMLHttpRequest.DONE) {\n          return;\n        }\n        if (ajax.status !== 200) {\n          _this.props.dispatch((0, _account.loadError)());\n        }\n        var payload = JSON.parse(ajax.response);\n        _this.props.dispatch((0, _account.loadSuccess)(payload.data));\n      };\n      ajax.send();\n      _this.props.dispatch((0, _account.loadStart)());\n    }, _this.parseSections = function (sections) {\n      sections.forEach(function (section) {\n\n        _this.props.dispatch((0, _section.addSection)({\n          id: section.id,\n          title: section.title,\n          intro: section.intro,\n          body: section.body,\n          imageUrl: section.imageUrl,\n          AccountId: section.AccountId\n        }));\n      });\n    }, _this.parseUpdates = function (updates) {\n      updates.forEach(function (update) {\n        _this.parseSections(update.Sections);\n        var sections = update.Sections.map(function (section) {\n          return section.id;\n        });\n        _this.props.dispatch((0, _update.addUpdate)({\n          id: update.id,\n          draft: update.draft,\n          sections: sections,\n          AccountId: update.AccountId\n        }));\n      });\n    }, _this.parseDigests = function (digests) {\n      digests.forEach(function (digest) {\n        _this.parseUpdates(digest.Updates);\n        var updates = digest.Updates.map(function (update) {\n          return update.id;\n        });\n        _this.props.dispatch((0, _digest.addDigest)({\n          id: digest.id,\n          readAt: digest.readAt,\n          sentAt: digest.sentAt,\n          updates: updates,\n          PersonId: digest.PersonId\n        }));\n      });\n    }, _this.fetchAllDigests = function () {\n      var ajax = new XMLHttpRequest();\n      ajax.open('GET', '/api/digests');\n      ajax.onreadystatechange = function () {\n        if (ajax.readyState != XMLHttpRequest.DONE) {\n          return;\n        }\n        if (ajax.status !== 200) {\n          _this.props.dispatch((0, _digest.loadError)());\n        }\n        var payload = JSON.parse(ajax.response).data;\n\n        _this.parseDigests(payload);\n        _this.props.dispatch((0, _digest.loadSuccess)());\n      };\n      ajax.send();\n      _this.props.dispatch((0, _digest.loadStart)());\n    }, _this.parseUpdates = function (updates) {\n      updates.forEach(function (update) {\n        _this.parseSections(update.Sections);\n        var sections = update.Sections.map(function (section) {\n          return section.id;\n        });\n        _this.props.dispatch((0, _update.addUpdate)({\n          id: update.id,\n          draft: update.draft,\n          sections: sections,\n          AccountId: update.AccountId\n        }));\n      });\n    }, _this.fetchAllUpdates = function () {\n      var ajax = new XMLHttpRequest();\n      ajax.open('GET', '/api/updates');\n      ajax.onreadystatechange = function () {\n        if (ajax.readyState != XMLHttpRequest.DONE) {\n          return;\n        }\n        if (ajax.status !== 200) {\n          _this.props.dispatch((0, _myUpdate.loadError)());\n        }\n        var payload = JSON.parse(ajax.response);\n\n        _this.parseMyUpdates(payload.updates);\n        _this.props.dispatch((0, _myUpdate.loadSuccess)());\n      };\n      ajax.send();\n      _this.props.dispatch((0, _myUpdate.loadStart)());\n    }, _this.parseMyUpdates = function (updates) {\n      updates.forEach(function (update) {\n        _this.parseSections(update.Sections);\n        var sections = update.Sections.map(function (section) {\n          return section.id;\n        });\n        _this.props.dispatch((0, _myUpdate.addUpdate)({\n          id: update.id,\n          draft: update.draft,\n          sections: sections,\n          AccountId: update.AccountId\n        }));\n      });\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(Navbar, [{\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      if (this.props.currentUser.get('loading')) {\n        return _react2.default.createElement(\n          'h1',\n          null,\n          'Loading...'\n        );\n      }\n\n      return _react2.default.createElement(\n        _reactMdl.Layout,\n        { fixedHeader: true, fixedDrawer: true },\n        _react2.default.createElement(\n          _reactMdl.Drawer,\n          { title: 'Kindred' },\n          _react2.default.createElement('img', { src: this.props.currentUser.get('avatarUrl'), height: '240px' }),\n          _react2.default.createElement(\n            'p',\n            null,\n            this.props.currentUser.get('firstName') + ' ' + this.props.currentUser.get('lastName')\n          ),\n          _react2.default.createElement(\n            _reactMdl.Navigation,\n            null,\n            _react2.default.createElement(\n              'a',\n              { onClick: function onClick() {\n                  _this2.props.dispatch((0, _reduxSimpleRouter.pushPath)('/'));\n                } },\n              'Dashboard'\n            ),\n            _react2.default.createElement(\n              'a',\n              { onClick: function onClick() {\n                  _this2.props.dispatch((0, _reduxSimpleRouter.pushPath)('/editor'));\n                } },\n              'Editor'\n            ),\n            _react2.default.createElement(\n              'a',\n              { onClick: function onClick() {\n                  _this2.props.dispatch((0, _reduxSimpleRouter.pushPath)('/updates'));\n                } },\n              'My Updates'\n            )\n          )\n        ),\n        _react2.default.createElement(\n          _reactMdl.Header,\n          { title: 'Kindred' },\n          _react2.default.createElement(_reactMdl.Textfield, {\n            label: 'Search',\n            expandable: true,\n            expandableIcon: 'search'\n          })\n        ),\n        _react2.default.createElement(\n          'main',\n          { className: 'mdl-layout__content' },\n          _react2.default.createElement(\n            'div',\n            { className: 'page-content' },\n            this.props.children\n          )\n        )\n      );\n    }\n  }]);\n\n  return Navbar;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    currentUser: state.currentUser,\n    loading: state.content.loading\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(Navbar);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"navbar.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/navbar.js\n ** module id = 242\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/navbar.js?");
 
 /***/ },
 /* 243 */
@@ -2318,85 +2318,85 @@
 /* 295 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _digestCard = __webpack_require__(296);\n\nvar _digestCard2 = _interopRequireDefault(_digestCard);\n\nvar _digest = __webpack_require__(297);\n\nvar _reactRedux = __webpack_require__(180);\n\nvar _reduxSimpleRouter = __webpack_require__(240);\n\nvar _digest2 = __webpack_require__(297);\n\nvar _section = __webpack_require__(298);\n\nvar _update = __webpack_require__(299);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar DigestList = (function (_React$Component) {\n  _inherits(DigestList, _React$Component);\n\n  function DigestList() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, DigestList);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(DigestList)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.componentDidMount = function () {\n      console.log('asd');\n      if (_this.props.digests.size === 0) {\n        _this.fetchAllDigests();\n      } else {\n        _this.props.digests.forEach(function (digest) {\n          _this.getImageUrl(digest.get('id'));\n        });\n      }\n    }, _this.readFullDigest = function (id) {\n      _this.props.dispatch((0, _reduxSimpleRouter.pushPath)('/digests/' + id));\n    }, _this.getImageUrl = function (id) {\n      function getRandomArrayElement(arr) {\n        return arr[Math.floor(Math.random() * arr.length)];\n      }\n\n      var imageUrl = undefined;\n\n      var digest = _this.props.digests.get(id);\n\n      // gets random image URL to display in update page, clean up this messyness\n      if (digest.get('imageUrl') === undefined) {\n        var randUpdate = _this.props.updates.get(getRandomArrayElement(digest.get('updates')));\n        var randSection = _this.props.sections.get(getRandomArrayElement(randUpdate.get('sections')));\n        imageUrl = randSection.get('imageUrl');\n\n        _this.props.dispatch((0, _digest.saveImageUrl)(digest.get('id'), imageUrl));\n      } else {\n        imageUrl = digest.get('imageUrl');\n      }\n      return imageUrl;\n    }, _this.parseSections = function (sections) {\n      sections.forEach(function (section) {\n\n        _this.props.dispatch((0, _section.addSection)({\n          id: section.id,\n          title: section.title,\n          intro: section.intro,\n          body: section.body,\n          imageUrl: section.imageUrl,\n          AccountId: section.AccountId\n        }));\n      });\n    }, _this.parseUpdates = function (updates) {\n      updates.forEach(function (update) {\n        _this.parseSections(update.Sections);\n        var sections = update.Sections.map(function (section) {\n          return section.id;\n        });\n        _this.props.dispatch((0, _update.addUpdate)({\n          id: update.id,\n          draft: update.draft,\n          sections: sections,\n          AccountId: update.AccountId\n        }));\n      });\n    }, _this.parseDigests = function (digests) {\n      digests.forEach(function (digest) {\n        _this.parseUpdates(digest.Updates);\n        var updates = digest.Updates.map(function (update) {\n          return update.id;\n        });\n        _this.props.dispatch((0, _digest2.addDigest)({\n          id: digest.id,\n          readAt: digest.readAt,\n          sentAt: digest.sentAt,\n          updates: updates,\n          PersonId: digest.PersonId\n        }));\n      });\n    }, _this.fetchAllDigests = function () {\n      var ajax = new XMLHttpRequest();\n      ajax.open('GET', '/api/digests');\n      ajax.onreadystatechange = function () {\n        if (ajax.readyState != XMLHttpRequest.DONE) {\n          return;\n        }\n        if (ajax.status !== 200) {\n          _this.props.dispatch((0, _digest2.loadError)());\n        }\n        var payload = JSON.parse(ajax.response).data;\n\n        _this.parseDigests(payload);\n        _this.props.dispatch((0, _digest2.loadSuccess)());\n      };\n      ajax.send();\n      _this.props.dispatch((0, _digest2.loadStart)());\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(DigestList, [{\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      var digests = [];\n\n      this.props.digests.forEach(function (digest) {\n\n        digests.push(_react2.default.createElement(_digestCard2.default, {\n          id: digest.get('id'),\n          key: digest.get('id'),\n          imageUrl: digest.get('imageUrl'),\n          publishedAt: digest.get('sentAt'),\n          readFull: _this2.readFullDigest\n        }));\n      });\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        digests\n      );\n    }\n  }]);\n\n  return DigestList;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    digests: state.content.digests,\n    updates: state.content.updates,\n    sections: state.content.sections\n  };\n}\nexports.default = (0, _reactRedux.connect)(mapState)(DigestList);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"digest-list.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/digest-list.js\n ** module id = 295\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/digest-list.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.saveImageUrl = saveImageUrl;\nexports.loadStart = loadStart;\nexports.loadSuccess = loadSuccess;\nexports.loadError = loadError;\nexports.addDigest = addDigest;\n\nvar _index = __webpack_require__(291);\n\nfunction saveImageUrl(digestId, imageUrl) {\n  return {\n    imageUrl: imageUrl,\n    id: digestId,\n    type: _index.DigestConstants.SAVE_IMAGE_URL\n  };\n}\n\nfunction loadStart() {\n  return {\n    type: _index.DigestConstants.LOAD_START\n  };\n}\n\nfunction loadSuccess() {\n  return {\n    type: _index.DigestConstants.LOAD_SUCCESS\n  };\n}\n\nfunction loadError() {\n  return {\n    type: _index.DigestConstants.LOAD_ERROR\n  };\n}\n\nfunction addDigest(payload) {\n  return {\n    id: payload.id,\n    readAt: payload.readAt,\n    sentAt: payload.sentAt,\n    updates: payload.updates,\n    PersonId: payload.PersonId,\n    type: _index.DigestConstants.ADD_DIGEST\n  };\n}\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"digest.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/actions/digest.js\n ** module id = 295\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/actions/digest.js?");
 
 /***/ },
 /* 296 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactMdl = __webpack_require__(243);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar DigestCard = (function (_React$Component) {\n  _inherits(DigestCard, _React$Component);\n\n  function DigestCard() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, DigestCard);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(DigestCard)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.readFull = function () {\n      _this.props.readFull(_this.props.id);\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(DigestCard, [{\n    key: 'render',\n    value: function render() {\n      var publishedAt = new Date(this.props.publishedAt);\n      var DateString = publishedAt.toDateString();\n      return _react2.default.createElement(\n        _reactMdl.Card,\n        { className: 'mdl-card-dash mdl-cell mdl-cell--4-col mdl-cell--4-col-tablet mdl-shadow--2dp',\n          style: { background: 'url(\\'' + this.props.imageUrl + '\\') center / cover no-repeat #46B6AC' },\n          onClick: this.readFull },\n        _react2.default.createElement(\n          _reactMdl.CardTitle,\n          null,\n          _react2.default.createElement(\n            'h2',\n            { className: 'mdl-card__title-text' },\n            DateString\n          )\n        )\n      );\n    }\n  }]);\n\n  return DigestCard;\n})(_react2.default.Component);\n\nexports.default = DigestCard;\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"digest-card.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/digest-card.js\n ** module id = 296\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/digest-card.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.readFull = readFull;\nexports.addSection = addSection;\n\nvar _index = __webpack_require__(291);\n\nfunction readFull(id) {\n  return {\n    type: _index.SectionConstants.TOGGLE_VIEW,\n    id: id\n  };\n}\n\nfunction addSection(section) {\n  return {\n    id: section.id,\n    title: section.title,\n    intro: section.intro,\n    body: section.body,\n    imageUrl: section.imageUrl,\n    comments: section.comments,\n    AccountId: section.AccountId,\n    type: _index.SectionConstants.ADD_SECTION\n  };\n}\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"section.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/actions/section.js\n ** module id = 296\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/actions/section.js?");
 
 /***/ },
 /* 297 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.saveImageUrl = saveImageUrl;\nexports.loadStart = loadStart;\nexports.loadSuccess = loadSuccess;\nexports.loadError = loadError;\nexports.addDigest = addDigest;\n\nvar _index = __webpack_require__(291);\n\nfunction saveImageUrl(digestId, imageUrl) {\n  return {\n    imageUrl: imageUrl,\n    id: digestId,\n    type: _index.DigestConstants.SAVE_IMAGE_URL\n  };\n}\n\nfunction loadStart() {\n  return {\n    type: _index.DigestConstants.LOAD_START\n  };\n}\n\nfunction loadSuccess() {\n  return {\n    type: _index.DigestConstants.LOAD_SUCCESS\n  };\n}\n\nfunction loadError() {\n  return {\n    type: _index.DigestConstants.LOAD_ERROR\n  };\n}\n\nfunction addDigest(payload) {\n  return {\n    id: payload.id,\n    readAt: payload.readAt,\n    sentAt: payload.sentAt,\n    updates: payload.updates,\n    PersonId: payload.PersonId,\n    type: _index.DigestConstants.ADD_DIGEST\n  };\n}\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"digest.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/actions/digest.js\n ** module id = 297\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/actions/digest.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.saveImageUrl = saveImageUrl;\nexports.loadStart = loadStart;\nexports.loadSuccess = loadSuccess;\nexports.loadError = loadError;\nexports.addUpdate = addUpdate;\n\nvar _index = __webpack_require__(291);\n\nfunction saveImageUrl(updateId, imageUrl) {\n  return {\n    imageUrl: imageUrl,\n    id: updateId,\n    type: _index.UpdateConstants.SAVE_IMAGE_URL\n  };\n}\n\nfunction loadStart() {\n  return {\n    type: _index.UpdateConstants.LOAD_START\n  };\n}\n\nfunction loadSuccess() {\n  return {\n    type: _index.UpdateConstants.LOAD_SUCCESS\n  };\n}\n\nfunction loadError() {\n  return {\n    type: _index.UpdateConstants.LOAD_ERROR\n  };\n}\n\nfunction addUpdate(update) {\n  return {\n    id: update.id,\n    draft: update.draft,\n    sections: update.sections,\n    AccountId: update.AccountId,\n    type: _index.UpdateConstants.ADD_UPDATE\n  };\n}\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"update.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/actions/update.js\n ** module id = 297\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/actions/update.js?");
 
 /***/ },
 /* 298 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.readFull = readFull;\nexports.addSection = addSection;\n\nvar _index = __webpack_require__(291);\n\nfunction readFull(id) {\n  return {\n    type: _index.SectionConstants.TOGGLE_VIEW,\n    id: id\n  };\n}\n\nfunction addSection(section) {\n  return {\n    id: section.id,\n    title: section.title,\n    intro: section.intro,\n    body: section.body,\n    imageUrl: section.imageUrl,\n    comments: section.comments,\n    AccountId: section.AccountId,\n    type: _index.SectionConstants.ADD_SECTION\n  };\n}\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"section.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/actions/section.js\n ** module id = 298\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/actions/section.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.saveImageUrl = saveImageUrl;\nexports.loadStart = loadStart;\nexports.loadSuccess = loadSuccess;\nexports.loadError = loadError;\nexports.addUpdate = addUpdate;\n\nvar _index = __webpack_require__(291);\n\nfunction saveImageUrl(updateId, imageUrl) {\n  return {\n    imageUrl: imageUrl,\n    id: updateId,\n    type: _index.MyUpdateConstants.SAVE_IMAGE_URL\n  };\n}\n\nfunction loadStart() {\n  return {\n    type: _index.MyUpdateConstants.LOAD_START\n  };\n}\n\nfunction loadSuccess() {\n  return {\n    type: _index.MyUpdateConstants.LOAD_SUCCESS\n  };\n}\n\nfunction loadError() {\n  return {\n    type: _index.MyUpdateConstants.LOAD_ERROR\n  };\n}\n\nfunction addUpdate(update) {\n  return {\n    id: update.id,\n    draft: update.draft,\n    sections: update.sections,\n    AccountId: update.AccountId,\n    type: _index.MyUpdateConstants.ADD_UPDATE\n  };\n}\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"my-update.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/actions/my-update.js\n ** module id = 298\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/actions/my-update.js?");
 
 /***/ },
 /* 299 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.saveImageUrl = saveImageUrl;\nexports.loadStart = loadStart;\nexports.loadSuccess = loadSuccess;\nexports.loadError = loadError;\nexports.addUpdate = addUpdate;\n\nvar _index = __webpack_require__(291);\n\nfunction saveImageUrl(updateId, imageUrl) {\n  return {\n    imageUrl: imageUrl,\n    id: updateId,\n    type: _index.UpdateConstants.SAVE_IMAGE_URL\n  };\n}\n\nfunction loadStart() {\n  return {\n    type: _index.UpdateConstants.LOAD_START\n  };\n}\n\nfunction loadSuccess() {\n  return {\n    type: _index.UpdateConstants.LOAD_SUCCESS\n  };\n}\n\nfunction loadError() {\n  return {\n    type: _index.UpdateConstants.LOAD_ERROR\n  };\n}\n\nfunction addUpdate(update) {\n  return {\n    id: update.id,\n    draft: update.draft,\n    sections: update.sections,\n    AccountId: update.AccountId,\n    type: _index.UpdateConstants.ADD_UPDATE\n  };\n}\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"update.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/actions/update.js\n ** module id = 299\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/actions/update.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _digestCard = __webpack_require__(300);\n\nvar _digestCard2 = _interopRequireDefault(_digestCard);\n\nvar _digest = __webpack_require__(295);\n\nvar _reactRedux = __webpack_require__(180);\n\nvar _reduxSimpleRouter = __webpack_require__(240);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar DigestList = (function (_React$Component) {\n  _inherits(DigestList, _React$Component);\n\n  function DigestList() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, DigestList);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(DigestList)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.componentDidMount = function () {\n      if (_this.props.digests.size !== 0) {\n        _this.props.digests.forEach(function (digest) {\n          _this.getImageUrl(digest.get('id'));\n        });\n      }\n    }, _this.readFullDigest = function (id) {\n      _this.props.dispatch((0, _reduxSimpleRouter.pushPath)('/digests/' + id));\n    }, _this.getImageUrl = function (id) {\n      function getRandomArrayElement(arr) {\n        return arr[Math.floor(Math.random() * arr.length)];\n      }\n\n      var imageUrl = undefined;\n\n      var digest = _this.props.digests.get(id);\n\n      // gets random image URL to display in update page, clean up this messyness\n      if (digest.get('imageUrl') === undefined) {\n        var randUpdate = _this.props.updates.get(getRandomArrayElement(digest.get('updates')));\n        var randSection = _this.props.sections.get(getRandomArrayElement(randUpdate.get('sections')));\n        imageUrl = randSection.get('imageUrl');\n\n        _this.props.dispatch((0, _digest.saveImageUrl)(digest.get('id'), imageUrl));\n      } else {\n        imageUrl = digest.get('imageUrl');\n      }\n      return imageUrl;\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(DigestList, [{\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      var digests = [];\n\n      this.props.digests.forEach(function (digest) {\n\n        digests.push(_react2.default.createElement(_digestCard2.default, {\n          id: digest.get('id'),\n          key: digest.get('id'),\n          imageUrl: digest.get('imageUrl'),\n          publishedAt: digest.get('sentAt'),\n          readFull: _this2.readFullDigest\n        }));\n      });\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        digests\n      );\n    }\n  }]);\n\n  return DigestList;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    digests: state.content.digests,\n    updates: state.content.updates,\n    sections: state.content.sections\n  };\n}\nexports.default = (0, _reactRedux.connect)(mapState)(DigestList);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"digest-list.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/digest-list.js\n ** module id = 299\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/digest-list.js?");
 
 /***/ },
 /* 300 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Editor = (function (_React$Component) {\n  _inherits(Editor, _React$Component);\n\n  function Editor() {\n    _classCallCheck(this, Editor);\n\n    return _possibleConstructorReturn(this, Object.getPrototypeOf(Editor).apply(this, arguments));\n  }\n\n  _createClass(Editor, [{\n    key: 'render',\n    value: function render() {\n      return _react2.default.createElement(\n        'h1',\n        null,\n        'Editor'\n      );\n    }\n  }]);\n\n  return Editor;\n})(_react2.default.Component);\n\nexports.default = Editor;\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"editor.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/editor.js\n ** module id = 300\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/editor.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactMdl = __webpack_require__(243);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar DigestCard = (function (_React$Component) {\n  _inherits(DigestCard, _React$Component);\n\n  function DigestCard() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, DigestCard);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(DigestCard)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.readFull = function () {\n      _this.props.readFull(_this.props.id);\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(DigestCard, [{\n    key: 'render',\n    value: function render() {\n      var publishedAt = new Date(this.props.publishedAt);\n      var DateString = publishedAt.toDateString();\n      return _react2.default.createElement(\n        _reactMdl.Card,\n        { className: 'mdl-card-dash mdl-cell mdl-cell--4-col mdl-cell--4-col-tablet mdl-shadow--2dp',\n          style: { background: 'url(\\'' + this.props.imageUrl + '\\') center / cover no-repeat #46B6AC' },\n          onClick: this.readFull },\n        _react2.default.createElement(\n          _reactMdl.CardTitle,\n          null,\n          _react2.default.createElement(\n            'h2',\n            { className: 'mdl-card__title-text' },\n            DateString\n          )\n        )\n      );\n    }\n  }]);\n\n  return DigestCard;\n})(_react2.default.Component);\n\nexports.default = DigestCard;\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"digest-card.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/digest-card.js\n ** module id = 300\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/digest-card.js?");
 
 /***/ },
 /* 301 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _sectionContainer = __webpack_require__(302);\n\nvar _sectionContainer2 = _interopRequireDefault(_sectionContainer);\n\nvar _reactRedux = __webpack_require__(180);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Digest = (function (_React$Component) {\n  _inherits(Digest, _React$Component);\n\n  function Digest() {\n    _classCallCheck(this, Digest);\n\n    return _possibleConstructorReturn(this, Object.getPrototypeOf(Digest).apply(this, arguments));\n  }\n\n  _createClass(Digest, [{\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      var currentDigest = this.props.digests.get(parseInt(this.props.id));\n      var sections = [];\n\n      currentDigest.get('updates').forEach(function (updateID) {\n        _this2.props.updates.get(updateID).get('sections').forEach(function (sectionId) {\n\n          sections.push(_react2.default.createElement(_sectionContainer2.default, {\n            id: sectionId,\n            key: sectionId\n          }));\n        });\n      });\n\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        sections\n      );\n    }\n  }]);\n\n  return Digest;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    digests: state.content.digests,\n    updates: state.content.updates,\n    sections: state.content.sections\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(Digest);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"digest.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/digest.js\n ** module id = 301\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/digest.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Editor = (function (_React$Component) {\n  _inherits(Editor, _React$Component);\n\n  function Editor() {\n    _classCallCheck(this, Editor);\n\n    return _possibleConstructorReturn(this, Object.getPrototypeOf(Editor).apply(this, arguments));\n  }\n\n  _createClass(Editor, [{\n    key: 'render',\n    value: function render() {\n      return _react2.default.createElement(\n        'h1',\n        null,\n        'Editor'\n      );\n    }\n  }]);\n\n  return Editor;\n})(_react2.default.Component);\n\nexports.default = Editor;\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"editor.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/editor.js\n ** module id = 301\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/editor.js?");
 
 /***/ },
 /* 302 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _section = __webpack_require__(303);\n\nvar _section2 = _interopRequireDefault(_section);\n\nvar _reactRedux = __webpack_require__(180);\n\nvar _section3 = __webpack_require__(298);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar SectionContainer = (function (_React$Component) {\n  _inherits(SectionContainer, _React$Component);\n\n  function SectionContainer() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, SectionContainer);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(SectionContainer)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.readFullSection = function (id) {\n      _this.props.dispatch((0, _section3.readFull)(id));\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(SectionContainer, [{\n    key: 'render',\n    value: function render() {\n      var id = parseInt(this.props.id);\n      var currentSection = this.props.sections.get(id);\n\n      return _react2.default.createElement(_section2.default, {\n        id: currentSection.get('id'),\n        imageUrl: currentSection.get('imageUrl'),\n        title: currentSection.get('title'),\n        intro: currentSection.get('intro'),\n        body: currentSection.get('body'),\n        fullView: currentSection.get('fullView'),\n        readFullSection: this.readFullSection\n      });\n    }\n  }]);\n\n  return SectionContainer;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    sections: state.content.sections\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(SectionContainer);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"section-container.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/section-container.js\n ** module id = 302\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/section-container.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _sectionContainer = __webpack_require__(303);\n\nvar _sectionContainer2 = _interopRequireDefault(_sectionContainer);\n\nvar _reactRedux = __webpack_require__(180);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Digest = (function (_React$Component) {\n  _inherits(Digest, _React$Component);\n\n  function Digest() {\n    _classCallCheck(this, Digest);\n\n    return _possibleConstructorReturn(this, Object.getPrototypeOf(Digest).apply(this, arguments));\n  }\n\n  _createClass(Digest, [{\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      var currentDigest = this.props.digests.get(parseInt(this.props.id));\n      var sections = [];\n\n      currentDigest.get('updates').forEach(function (updateID) {\n        _this2.props.updates.get(updateID).get('sections').forEach(function (sectionId) {\n\n          sections.push(_react2.default.createElement(_sectionContainer2.default, {\n            id: sectionId,\n            key: sectionId\n          }));\n        });\n      });\n\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        sections\n      );\n    }\n  }]);\n\n  return Digest;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    digests: state.content.digests,\n    updates: state.content.updates,\n    sections: state.content.sections\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(Digest);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"digest.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/digest.js\n ** module id = 302\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/digest.js?");
 
 /***/ },
 /* 303 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactMdl = __webpack_require__(243);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Section = (function (_React$Component) {\n  _inherits(Section, _React$Component);\n\n  function Section() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, Section);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(Section)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.readFull = function () {\n      _this.props.readFullSection(_this.props.id);\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(Section, [{\n    key: 'render',\n    value: function render() {\n      return _react2.default.createElement(\n        _reactMdl.Card,\n        { className: 'mdl-card-digest mdl-cell mdl-cell--12-col mdl-cell--12-col-tablet mdl-shadow--2dp' },\n        _react2.default.createElement(\n          _reactMdl.CardTitle,\n          {\n            style: { background: 'url(\\'' + this.props.imageUrl + '\\') center / cover no-repeat #46B6AC' } },\n          this.props.title\n        ),\n        _react2.default.createElement(\n          _reactMdl.CardText,\n          null,\n          this.props.fullView ? this.props.intro + ' ' + this.props.body : this.props.intro,\n          _react2.default.createElement(\n            'div',\n            { className: 'mdl-typography--text-center' },\n            _react2.default.createElement(\n              'a',\n              { onClick: this.readFull },\n              this.props.fullView ? \"Less\" : \"Read More\"\n            )\n          )\n        )\n      );\n    }\n  }]);\n\n  return Section;\n})(_react2.default.Component);\n\nexports.default = Section;\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"section.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/section.js\n ** module id = 303\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/section.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _section = __webpack_require__(304);\n\nvar _section2 = _interopRequireDefault(_section);\n\nvar _reactRedux = __webpack_require__(180);\n\nvar _section3 = __webpack_require__(296);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar SectionContainer = (function (_React$Component) {\n  _inherits(SectionContainer, _React$Component);\n\n  function SectionContainer() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, SectionContainer);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(SectionContainer)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.readFullSection = function (id) {\n      _this.props.dispatch((0, _section3.readFull)(id));\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(SectionContainer, [{\n    key: 'render',\n    value: function render() {\n      var id = parseInt(this.props.id);\n      var currentSection = this.props.sections.get(id);\n\n      return _react2.default.createElement(_section2.default, {\n        id: currentSection.get('id'),\n        imageUrl: currentSection.get('imageUrl'),\n        title: currentSection.get('title'),\n        intro: currentSection.get('intro'),\n        body: currentSection.get('body'),\n        fullView: currentSection.get('fullView'),\n        readFullSection: this.readFullSection\n      });\n    }\n  }]);\n\n  return SectionContainer;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    sections: state.content.sections\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(SectionContainer);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"section-container.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/section-container.js\n ** module id = 303\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/section-container.js?");
 
 /***/ },
 /* 304 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _sectionContainer = __webpack_require__(302);\n\nvar _sectionContainer2 = _interopRequireDefault(_sectionContainer);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Update = (function (_React$Component) {\n  _inherits(Update, _React$Component);\n\n  function Update() {\n    _classCallCheck(this, Update);\n\n    return _possibleConstructorReturn(this, Object.getPrototypeOf(Update).apply(this, arguments));\n  }\n\n  _createClass(Update, [{\n    key: 'render',\n    value: function render() {\n      var sections = [];\n\n      this.props.sections.forEach(function (sectionId) {\n        sections.push(_react2.default.createElement(_sectionContainer2.default, {\n          id: sectionId,\n          key: sectionId\n        }));\n      });\n\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        sections\n      );\n    }\n  }]);\n\n  return Update;\n})(_react2.default.Component);\n\nexports.default = Update;\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"update.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/update.js\n ** module id = 304\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/update.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _reactMdl = __webpack_require__(243);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Section = (function (_React$Component) {\n  _inherits(Section, _React$Component);\n\n  function Section() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, Section);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(Section)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.readFull = function () {\n      _this.props.readFullSection(_this.props.id);\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(Section, [{\n    key: 'render',\n    value: function render() {\n      return _react2.default.createElement(\n        _reactMdl.Card,\n        { className: 'mdl-card-digest mdl-cell mdl-cell--12-col mdl-cell--12-col-tablet mdl-shadow--2dp' },\n        _react2.default.createElement(\n          _reactMdl.CardTitle,\n          {\n            style: { background: 'url(\\'' + this.props.imageUrl + '\\') center / cover no-repeat #46B6AC' } },\n          this.props.title\n        ),\n        _react2.default.createElement(\n          _reactMdl.CardText,\n          null,\n          this.props.fullView ? this.props.intro + ' ' + this.props.body : this.props.intro,\n          _react2.default.createElement(\n            'div',\n            { className: 'mdl-typography--text-center' },\n            _react2.default.createElement(\n              'a',\n              { onClick: this.readFull },\n              this.props.fullView ? \"Less\" : \"Read More\"\n            )\n          )\n        )\n      );\n    }\n  }]);\n\n  return Section;\n})(_react2.default.Component);\n\nexports.default = Section;\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"section.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/section.js\n ** module id = 304\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/section.js?");
 
 /***/ },
 /* 305 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _digestCard = __webpack_require__(296);\n\nvar _digestCard2 = _interopRequireDefault(_digestCard);\n\nvar _reactRedux = __webpack_require__(180);\n\nvar _reduxSimpleRouter = __webpack_require__(240);\n\nvar _myUpdate = __webpack_require__(306);\n\nvar _section = __webpack_require__(298);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar UpdateList = (function (_React$Component) {\n  _inherits(UpdateList, _React$Component);\n\n  function UpdateList() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, UpdateList);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(UpdateList)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.componentDidMount = function () {\n      console.log('asdff');\n      if (_this.props.updates.size === 0) {\n        _this.fetchAllUpdates();\n      } else {\n        _this.props.updates.forEach(function (update) {\n          _this.getImageUrl(update.get('id'));\n        });\n      }\n    }, _this.parseSections = function (sections) {\n      sections.forEach(function (section) {\n        var comments = section.Comments.map(function (comment) {\n          return comment.id;\n        });\n\n        _this.props.dispatch((0, _section.addSection)({\n          id: section.id,\n          title: section.title,\n          intro: section.intro,\n          body: section.body,\n          imageUrl: section.imageUrl,\n          comments: comments,\n          AccountId: section.AccountId\n        }));\n      });\n    }, _this.parseUpdates = function (updates) {\n      updates.forEach(function (update) {\n        _this.parseSections(update.Sections);\n        var sections = update.Sections.map(function (section) {\n          return section.id;\n        });\n        _this.props.dispatch((0, _myUpdate.addUpdate)({\n          id: update.id,\n          draft: update.draft,\n          sections: sections,\n          AccountId: update.AccountId\n        }));\n      });\n    }, _this.fetchAllUpdates = function () {\n      var ajax = new XMLHttpRequest();\n      ajax.open('GET', '/api/updates');\n      ajax.onreadystatechange = function () {\n        if (ajax.readyState != XMLHttpRequest.DONE) {\n          return;\n        }\n        if (ajax.status !== 200) {\n          _this.props.dispatch((0, _myUpdate.loadError)());\n        }\n        var payload = JSON.parse(ajax.response);\n\n        _this.parseUpdates(payload.updates);\n        _this.props.dispatch((0, _myUpdate.loadSuccess)());\n      };\n      ajax.send();\n      _this.props.dispatch((0, _myUpdate.loadStart)());\n    }, _this.readFullUpdate = function (id) {\n      _this.props.dispatch((0, _reduxSimpleRouter.pushPath)('/updates/' + id));\n    }, _this.getImageUrl = function (id) {\n      function getRandomArrayElement(arr) {\n        return arr[Math.floor(Math.random() * arr.length)];\n      }\n\n      var imageUrl = undefined;\n      var update = _this.props.updates.get(id);\n      // gets random image URL to display in update page, clean up this messyness\n      if (update.get('imageUrl') === undefined) {\n        var randSection = _this.props.sections.get(getRandomArrayElement(update.get('sections')));\n        imageUrl = randSection.get('imageUrl');\n\n        _this.props.dispatch((0, _myUpdate.saveImageUrl)(update.get('id'), imageUrl));\n      } else {\n        imageUrl = update.get('imageUrl');\n      }\n      return imageUrl;\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(UpdateList, [{\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      var updates = [];\n      this.props.updates.forEach(function (update) {\n        updates.push(_react2.default.createElement(_digestCard2.default, {\n          id: update.get('id'),\n          key: update.get('id'),\n          publishedAt: update.get('publishedAt'),\n          imageUrl: update.get('imageUrl'),\n          readFull: _this2.readFullUpdate\n        }));\n      });\n\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        updates\n      );\n    }\n  }]);\n\n  return UpdateList;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    updates: state.content.myUpdates,\n    sections: state.content.sections\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(UpdateList);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"update-list.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/update-list.js\n ** module id = 305\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/update-list.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _sectionContainer = __webpack_require__(303);\n\nvar _sectionContainer2 = _interopRequireDefault(_sectionContainer);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar Update = (function (_React$Component) {\n  _inherits(Update, _React$Component);\n\n  function Update() {\n    _classCallCheck(this, Update);\n\n    return _possibleConstructorReturn(this, Object.getPrototypeOf(Update).apply(this, arguments));\n  }\n\n  _createClass(Update, [{\n    key: 'render',\n    value: function render() {\n      var sections = [];\n\n      this.props.sections.forEach(function (sectionId) {\n        sections.push(_react2.default.createElement(_sectionContainer2.default, {\n          id: sectionId,\n          key: sectionId\n        }));\n      });\n\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        sections\n      );\n    }\n  }]);\n\n  return Update;\n})(_react2.default.Component);\n\nexports.default = Update;\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"update.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/update.js\n ** module id = 305\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/update.js?");
 
 /***/ },
 /* 306 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\nexports.saveImageUrl = saveImageUrl;\nexports.loadStart = loadStart;\nexports.loadSuccess = loadSuccess;\nexports.loadError = loadError;\nexports.addUpdate = addUpdate;\n\nvar _index = __webpack_require__(291);\n\nfunction saveImageUrl(updateId, imageUrl) {\n  return {\n    imageUrl: imageUrl,\n    id: updateId,\n    type: _index.MyUpdateConstants.SAVE_IMAGE_URL\n  };\n}\n\nfunction loadStart() {\n  return {\n    type: _index.MyUpdateConstants.LOAD_START\n  };\n}\n\nfunction loadSuccess() {\n  return {\n    type: _index.MyUpdateConstants.LOAD_SUCCESS\n  };\n}\n\nfunction loadError() {\n  return {\n    type: _index.MyUpdateConstants.LOAD_ERROR\n  };\n}\n\nfunction addUpdate(update) {\n  return {\n    id: update.id,\n    draft: update.draft,\n    sections: update.sections,\n    AccountId: update.AccountId,\n    type: _index.MyUpdateConstants.ADD_UPDATE\n  };\n}\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"my-update.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/actions/my-update.js\n ** module id = 306\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/actions/my-update.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _digestCard = __webpack_require__(300);\n\nvar _digestCard2 = _interopRequireDefault(_digestCard);\n\nvar _reactRedux = __webpack_require__(180);\n\nvar _reduxSimpleRouter = __webpack_require__(240);\n\nvar _myUpdate = __webpack_require__(298);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar UpdateList = (function (_React$Component) {\n  _inherits(UpdateList, _React$Component);\n\n  function UpdateList() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, UpdateList);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(UpdateList)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.componentDidMount = function () {\n      if (_this.props.updates.size === 0) {\n        _this.fetchAllUpdates();\n      } else {\n        _this.props.updates.forEach(function (update) {\n          _this.getImageUrl(update.get('id'));\n        });\n      }\n    }, _this.readFullUpdate = function (id) {\n      _this.props.dispatch((0, _reduxSimpleRouter.pushPath)('/updates/' + id));\n    }, _this.getImageUrl = function (id) {\n      function getRandomArrayElement(arr) {\n        return arr[Math.floor(Math.random() * arr.length)];\n      }\n\n      var imageUrl = undefined;\n      var update = _this.props.updates.get(id);\n      // gets random image URL to display in update page, clean up this messyness\n      if (update.get('imageUrl') === undefined) {\n        var randSection = _this.props.sections.get(getRandomArrayElement(update.get('sections')));\n        imageUrl = randSection.get('imageUrl');\n\n        _this.props.dispatch((0, _myUpdate.saveImageUrl)(update.get('id'), imageUrl));\n      } else {\n        imageUrl = update.get('imageUrl');\n      }\n      return imageUrl;\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(UpdateList, [{\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      var updates = [];\n      this.props.updates.forEach(function (update) {\n        updates.push(_react2.default.createElement(_digestCard2.default, {\n          id: update.get('id'),\n          key: update.get('id'),\n          publishedAt: update.get('publishedAt'),\n          imageUrl: update.get('imageUrl'),\n          readFull: _this2.readFullUpdate\n        }));\n      });\n\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        updates\n      );\n    }\n  }]);\n\n  return UpdateList;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    updates: state.content.myUpdates,\n    sections: state.content.sections\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(UpdateList);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"update-list.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/update-list.js\n ** module id = 306\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/update-list.js?");
 
 /***/ },
 /* 307 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _update = __webpack_require__(304);\n\nvar _update2 = _interopRequireDefault(_update);\n\nvar _reactRedux = __webpack_require__(180);\n\nvar _update3 = __webpack_require__(299);\n\nvar _section = __webpack_require__(298);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar UpdateContainer = (function (_React$Component) {\n  _inherits(UpdateContainer, _React$Component);\n\n  function UpdateContainer() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, UpdateContainer);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(UpdateContainer)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.componentDidMount = function () {\n      if (_this.props.updates.get(parseInt(_this.props.params.id))) {\n        return;\n      }\n      _this.fetchUpdate();\n    }, _this.parseSections = function (sections) {\n      sections.forEach(function (section) {\n        var comments = section.Comments.map(function (comment) {\n          return comment.id;\n        });\n\n        _this.props.dispatch((0, _section.addSection)({\n          id: section.id,\n          title: section.title,\n          intro: section.intro,\n          body: section.body,\n          imageUrl: section.imageUrl,\n          comments: comments,\n          AccountId: section.AccountId\n        }));\n      });\n    }, _this.fetchUpdate = function () {\n      var ajax = new XMLHttpRequest();\n      ajax.open('GET', '/api/updates/' + _this.props.params.id);\n      ajax.onreadystatechange = function () {\n        if (ajax.readyState != XMLHttpRequest.DONE) {\n          return;\n        }\n        if (ajax.status !== 200) {\n          _this.props.dispatch((0, _update3.loadError)());\n        }\n        var payload = JSON.parse(ajax.response).data;\n\n        _this.parseSections(payload.Sections);\n\n        _this.props.dispatch((0, _update3.addUpdate)({\n          id: payload.id,\n          draft: payload.draft,\n          AccountId: payload.AccountId,\n          sections: payload.Sections.map(function (section) {\n            return section.id;\n          })\n        }));\n\n        _this.props.dispatch((0, _update3.loadSuccess)());\n      };\n      ajax.send();\n      _this.props.dispatch((0, _update3.loadStart)());\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(UpdateContainer, [{\n    key: 'render',\n    value: function render() {\n      var currentUpdate = this.props.updates.get(parseInt(this.props.params.id));\n      if (currentUpdate === undefined) {\n        return _react2.default.createElement(\n          'h1',\n          null,\n          'Loading'\n        );\n      }\n      var sections = currentUpdate.get('sections');\n\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        _react2.default.createElement(_update2.default, { sections: sections })\n      );\n    }\n  }]);\n\n  return UpdateContainer;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    updates: state.content.updates\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(UpdateContainer);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"update-container.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/update-container.js\n ** module id = 307\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/update-container.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _update = __webpack_require__(305);\n\nvar _update2 = _interopRequireDefault(_update);\n\nvar _reactRedux = __webpack_require__(180);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar UpdateContainer = (function (_React$Component) {\n  _inherits(UpdateContainer, _React$Component);\n\n  function UpdateContainer() {\n    _classCallCheck(this, UpdateContainer);\n\n    return _possibleConstructorReturn(this, Object.getPrototypeOf(UpdateContainer).apply(this, arguments));\n  }\n\n  _createClass(UpdateContainer, [{\n    key: 'render',\n    value: function render() {\n      var currentUpdate = this.props.updates.get(parseInt(this.props.params.id));\n      if (currentUpdate === undefined) {\n        return _react2.default.createElement(\n          'h1',\n          null,\n          'Loading'\n        );\n      }\n      var sections = currentUpdate.get('sections');\n\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        _react2.default.createElement(_update2.default, { sections: sections })\n      );\n    }\n  }]);\n\n  return UpdateContainer;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    updates: state.content.updates\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(UpdateContainer);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"update-container.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/update-container.js\n ** module id = 307\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/update-container.js?");
 
 /***/ },
 /* 308 */
 /***/ function(module, exports, __webpack_require__) {
 
-	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _digest = __webpack_require__(301);\n\nvar _digest2 = _interopRequireDefault(_digest);\n\nvar _reactRedux = __webpack_require__(180);\n\nvar _digest3 = __webpack_require__(297);\n\nvar _section = __webpack_require__(298);\n\nvar _update = __webpack_require__(299);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar DigestContainer = (function (_React$Component) {\n  _inherits(DigestContainer, _React$Component);\n\n  function DigestContainer() {\n    var _Object$getPrototypeO;\n\n    var _temp, _this, _ret;\n\n    _classCallCheck(this, DigestContainer);\n\n    for (var _len = arguments.length, args = Array(_len), _key = 0; _key < _len; _key++) {\n      args[_key] = arguments[_key];\n    }\n\n    return _ret = (_temp = (_this = _possibleConstructorReturn(this, (_Object$getPrototypeO = Object.getPrototypeOf(DigestContainer)).call.apply(_Object$getPrototypeO, [this].concat(args))), _this), _this.componentDidMount = function () {\n      if (_this.props.digests.get(parseInt(_this.props.params.id))) {\n        return;\n      }\n      _this.fetchDigest();\n    }, _this.parseSections = function (sections) {\n      sections.forEach(function (section) {\n\n        _this.props.dispatch((0, _section.addSection)({\n          id: section.id,\n          title: section.title,\n          intro: section.intro,\n          body: section.body,\n          imageUrl: section.imageUrl,\n          AccountId: section.AccountId\n        }));\n      });\n    }, _this.parseUpdates = function (updates) {\n      updates.forEach(function (update) {\n        _this.parseSections(update.Sections);\n        var sections = update.Sections.map(function (section) {\n          return section.id;\n        });\n        _this.props.dispatch((0, _update.addUpdate)({\n          id: update.id,\n          draft: update.draft,\n          sections: sections,\n          AccountId: update.AccountId\n        }));\n      });\n    }, _this.parseDigests = function (digests) {\n      digests.forEach(function (digest) {\n        _this.parseUpdates(digest.Updates);\n        var updates = digest.Updates.map(function (update) {\n          return update.id;\n        });\n        _this.props.dispatch((0, _digest3.addDigest)({\n          id: digest.id,\n          readAt: digest.readAt,\n          sentAt: digest.sentAt,\n          updates: updates,\n          PersonId: digest.PersonId\n        }));\n      });\n    }, _this.fetchDigest = function () {\n      var ajax = new XMLHttpRequest();\n      ajax.open('GET', '/api/digests/' + _this.props.params.id);\n      ajax.onreadystatechange = function () {\n        if (ajax.readyState != XMLHttpRequest.DONE) {\n          return;\n        }\n        if (ajax.status !== 200) {\n          _this.props.dispatch((0, _digest3.loadError)());\n        }\n        var payload = JSON.parse(ajax.response).data;\n\n        _this.parseDigests([payload]);\n        _this.props.dispatch((0, _digest3.loadSuccess)());\n      };\n      ajax.send();\n      _this.props.dispatch((0, _digest3.loadStart)());\n    }, _temp), _possibleConstructorReturn(_this, _ret);\n  }\n\n  _createClass(DigestContainer, [{\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      var currentDigest = this.props.digests.get(parseInt(this.props.params.id));\n      if (currentDigest === undefined) {\n        return _react2.default.createElement(\n          'h1',\n          null,\n          'Loading'\n        );\n      }\n      var currentSections = [];\n\n      currentDigest.get('updates').forEach(function (updateID) {\n        _this2.props.updates.get(updateID).get('sections').forEach(function (sectionId) {\n          currentSections.push(sectionId);\n        });\n      });\n\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        _react2.default.createElement(_digest2.default, {\n          id: currentDigest.get('id'),\n          sections: currentSections\n        })\n      );\n    }\n  }]);\n\n  return DigestContainer;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    digests: state.content.digests,\n    updates: state.content.updates,\n    sections: state.content.sections\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(DigestContainer);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"digest-container.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/digest-container.js\n ** module id = 308\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/digest-container.js?");
+	eval("/* WEBPACK VAR INJECTION */(function(module) {/* REACT HOT LOADER */ if (true) { (function () { var ReactHotAPI = __webpack_require__(3), RootInstanceProvider = __webpack_require__(11), ReactMount = __webpack_require__(13), React = __webpack_require__(66); module.makeHot = module.hot.data ? module.hot.data.makeHot : ReactHotAPI(function () { return RootInstanceProvider.getRootInstances(ReactMount); }, React); })(); } try { (function () {\n\n'use strict';\n\nvar _createClass = (function () { function defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } } return function (Constructor, protoProps, staticProps) { if (protoProps) defineProperties(Constructor.prototype, protoProps); if (staticProps) defineProperties(Constructor, staticProps); return Constructor; }; })();\n\nObject.defineProperty(exports, \"__esModule\", {\n  value: true\n});\n\nvar _react = __webpack_require__(66);\n\nvar _react2 = _interopRequireDefault(_react);\n\nvar _digest = __webpack_require__(302);\n\nvar _digest2 = _interopRequireDefault(_digest);\n\nvar _reactRedux = __webpack_require__(180);\n\nfunction _interopRequireDefault(obj) { return obj && obj.__esModule ? obj : { default: obj }; }\n\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _possibleConstructorReturn(self, call) { if (!self) { throw new ReferenceError(\"this hasn't been initialised - super() hasn't been called\"); } return call && (typeof call === \"object\" || typeof call === \"function\") ? call : self; }\n\nfunction _inherits(subClass, superClass) { if (typeof superClass !== \"function\" && superClass !== null) { throw new TypeError(\"Super expression must either be null or a function, not \" + typeof superClass); } subClass.prototype = Object.create(superClass && superClass.prototype, { constructor: { value: subClass, enumerable: false, writable: true, configurable: true } }); if (superClass) Object.setPrototypeOf ? Object.setPrototypeOf(subClass, superClass) : subClass.__proto__ = superClass; }\n\nvar DigestContainer = (function (_React$Component) {\n  _inherits(DigestContainer, _React$Component);\n\n  function DigestContainer() {\n    _classCallCheck(this, DigestContainer);\n\n    return _possibleConstructorReturn(this, Object.getPrototypeOf(DigestContainer).apply(this, arguments));\n  }\n\n  _createClass(DigestContainer, [{\n    key: 'render',\n    value: function render() {\n      var _this2 = this;\n\n      var currentDigest = this.props.digests.get(parseInt(this.props.params.id));\n      if (currentDigest === undefined) {\n        return _react2.default.createElement(\n          'h1',\n          null,\n          'Loading'\n        );\n      }\n      var currentSections = [];\n\n      currentDigest.get('updates').forEach(function (updateID) {\n        _this2.props.updates.get(updateID).get('sections').forEach(function (sectionId) {\n          currentSections.push(sectionId);\n        });\n      });\n\n      return _react2.default.createElement(\n        'div',\n        { className: 'mdl-grid' },\n        _react2.default.createElement(_digest2.default, {\n          id: currentDigest.get('id'),\n          sections: currentSections\n        })\n      );\n    }\n  }]);\n\n  return DigestContainer;\n})(_react2.default.Component);\n\nfunction mapState(state) {\n  return {\n    digests: state.content.digests,\n    updates: state.content.updates,\n    sections: state.content.sections\n  };\n}\n\nexports.default = (0, _reactRedux.connect)(mapState)(DigestContainer);\n\n/* REACT HOT LOADER */ }).call(this); } finally { if (true) { (function () { var foundReactClasses = module.hot.data && module.hot.data.foundReactClasses || false; if (module.exports && module.makeHot) { var makeExportsHot = __webpack_require__(292); if (makeExportsHot(module, __webpack_require__(66))) { foundReactClasses = true; } var shouldAcceptModule = true && foundReactClasses; if (shouldAcceptModule) { module.hot.accept(function (err) { if (err) { console.error(\"Cannot not apply hot update to \" + \"digest-container.js\" + \": \" + err.message); } }); } } module.hot.dispose(function (data) { data.makeHot = module.makeHot; data.foundReactClasses = foundReactClasses; }); })(); } }\n/* WEBPACK VAR INJECTION */}.call(exports, __webpack_require__(2)(module)))\n\n/*****************\n ** WEBPACK FOOTER\n ** ./src/components/digest-container.js\n ** module id = 308\n ** module chunks = 0\n **/\n//# sourceURL=webpack:///./src/components/digest-container.js?");
 
 /***/ },
 /* 309 */
