+++ src/components/digest-list.js
@@ -6,17 +6,10 @@ import { saveImageUrl } from '../actions/digest'
 import { connect } from 'react-redux'
 import { pushPath } from 'redux-simple-router'
 
-import { loadStart, loadError, loadSuccess, addDigest } from '../actions/digest.js'
-import { addSection } from '../actions/section.js'
-import { addUpdate } from '../actions/update.js'
-
-
 class DigestList extends React.Component {
 
   componentDidMount = () => {
-    if ( this.props.digests.size === 0 ){
-      this.fetchAllDigests()
-    } else {
+    if ( this.props.digests.size !== 0 ){
       this.props.digests.forEach( (digest) =>{
         this.getImageUrl(digest.get('id'))
       })
@@ -49,69 +42,6 @@ class DigestList extends React.Component {
     return imageUrl
   }
 
-  parseSections = (sections) => {
-    sections.forEach( (section) => {
-
-      this.props.dispatch(addSection({
-        id: section.id,
-        title: section.title,
-        intro: section.intro,
-        body: section.body,
-        imageUrl: section.imageUrl,
-        AccountId: section.AccountId
-      }))
-    })
-  }
-
-  parseUpdates = (updates) => {
-    updates.forEach( (update) => {
-      this.parseSections(update.Sections)
-      let sections = update.Sections.map( (section) => {
-        return section.id
-      })
-      this.props.dispatch(addUpdate({
-        id: update.id,
-        draft: update.draft,
-        sections: sections,
-        AccountId: update.AccountId
-      }))
-    })
-  }
-
-  parseDigests = (digests) => {
-    digests.forEach( (digest) => {
-      this.parseUpdates(digest.Updates)
-      let updates = digest.Updates.map( (update) => {
-        return update.id
-      })
-      this.props.dispatch(addDigest({
-        id: digest.id,
-        readAt: digest.readAt,
-        sentAt: digest.sentAt,
-        updates: updates,
-        PersonId: digest.PersonId
-      }))
-    })
-  }
-
-  fetchAllDigests = () => {
-    let ajax = new XMLHttpRequest()
-    ajax.open('GET', '/api/digests')
-    ajax.onreadystatechange = () => {
-      if ( ajax.readyState != XMLHttpRequest.DONE ) {
-        return
-      }
-      if ( ajax.status !== 200 ) {
-        this.props.dispatch(loadError())
-      }
-      let payload = JSON.parse(ajax.response).data
-      
-      this.parseDigests(payload)
-      this.props.dispatch(loadSuccess())
-    }
-    ajax.send()
-    this.props.dispatch(loadStart())
-  }
 
   render() {
     let digests = []
